revert to a vphi equation instead of a psi equation
Slightly enhance the stability. But the integral formulation still blows up
very easily in the QG set of equations. This is likely because of the boundary
conditions of the 4th order operator (at the outer boundary we have hard time
to get dpsi/dr=0 since we regularise by (ro^2-r^2). I'm not sure there's a way
to save it or if maybe a psi/omega formulation with a 2*n_r_max block band
matrix would help.