preliminary support for full disk and variable m
For now works fine with an external fixed temperature but fails with Neumann
external BC and FD. Not sure where the problem comes from, but this likely also
occurs in the annulus configuration. Tests with Cheb also work fine, though with
quite a small timestep size with the GL grid.